Improve security by sanitizing and validating user inputs
This commit enhances the security of the front-end class by adding input sanitization and validation to prevent possible attacks. It replaces direct usage of $_REQUEST and $_SERVER with wp_unslash and sanitize_text_field. In addition, filter_var is used to validate IP addresses before usage, significantly increasing the robustness of the code against potential security threats.

And fix  get_ip not working